adding support for multiple outputs
specify 'sections=' in the [install] section, then list out the
extra sections you want to process. Each will be treated like it is
a separate [install] header. Values expected to remain the same for all
will still be pulled from [install], but of course may be specified in
[DEFAULT]. Values expected to commonly change will be referenced from
the individually named section (and of course back filled from [DEFAULT])

Also adds the 'exclude_extensions' option, which lets you exclude individual
extensions from dependency resolution as well as inclusion in the bundled image.

phone-tester.cfg is an example of this usage