update recursive directory search
Need to take into account having paths for multiple versions and architectures of TC in the search directory

So try adding in support for the TC mirror file structure, and removing other arches
and for moving other versions to the back of the search order. Not clear if we should try to
sort the other versions in the list or not.